Did a few modifications
