--- Birthday Cake.py
@@ -1,34 +1,0 @@
-
-import math
-import os
-import random
-import re
-import sys
-
-#
-# Complete the 'birthdayCakeCandles' function below.
-#
-# The function is expected to return an INTEGER.
-# The function accepts INTEGER_ARRAY candles as parameter.
-#
-
-def birthdayCakeCandles(candles):
-    sum=0
-    max_no = max(candles)
-    for i in range(len(candles)):
-        if(candles[i]== max_no):
-            sum+=1
-    return sum
-   
-if __name__ == '__main__':
-    fptr = open(os.environ['OUTPUT_PATH'], 'w')
-
-    candles_count = int(input().strip())
-
-    candles = list(map(int, input().rstrip().split()))
-
-    result = birthdayCakeCandles(candles)
-
-    fptr.write(str(result) + '\n')
-
-    fptr.close()
