+++ README.md
@@ -1,4 +1,8 @@
 # hacker-Rank-Problems
 
-i have been trying to solve alogrithms questions provided by haker Rank 
-all answers are in Pyhton
+I have been trying to solve alogrithms questions provided by haker Rank 
+all answers are in Pyhton. Check the hacker-Rank-problems-python directory for this
+
+And also I am trying to solve the problem solving part provided by hacker Rank 
+all answers are in C++. Check the hacker-Rank-problems-cpp directory for this.
+This directory is still under process, today's date is 18th April, 2021.
