+++ hacker-Rank-problems-python/A very big sum.py
@@ -1,0 +1,28 @@
+#!/bin/python3
+
+import math
+import os
+import random
+import re
+import sys
+
+# Complete the aVeryBigSum function below.
+def aVeryBigSum(ar,ar_count):
+    sum = 0
+    for i in range(ar_count):
+        sum = sum + ar[i]
+    return sum     
+
+
+if __name__ == '__main__':
+    fptr = open(os.environ['OUTPUT_PATH'], 'w')
+
+    ar_count = int(input())
+
+    ar = list(map(int, input().rstrip().split()))
+
+    result = aVeryBigSum(ar,ar_count)
+
+    fptr.write(str(result) + '\n')
+
+    fptr.close()
