+++ hacker-Rank-problems-python/diagnoal difference.py
@@ -1,0 +1,43 @@
+#!/bin/python3
+
+import math
+import os
+import random
+import re
+import sys
+
+#
+# Complete the 'diagonalDifference' function below.
+#
+# The function is expected to return an INTEGER.
+# The function accepts 2D_INTEGER_ARRAY arr as parameter.
+#
+
+def diagonalDifference(arr,n):
+    # Write your code here
+    d1=0
+    d2 = 0
+    for x in range(0,n):
+        for y in range(0,n):
+            if x == y: 
+                d1 += arr[x][y] 
+            if x == n-y-1: 
+                d2 += arr[x][y] 
+    return abs(d1 - d2)            
+            
+
+if __name__ == '__main__':
+    fptr = open(os.environ['OUTPUT_PATH'], 'w')
+
+    n = int(input().strip())
+
+    arr = []
+
+    for _ in range(n):
+        arr.append(list(map(int, input().rstrip().split())))
+
+    result = diagonalDifference(arr,n)
+
+    fptr.write(str(result) + '\n')
+
+    fptr.close()
