+++ hacker-Rank-problems-python/solve me first.py
@@ -1,0 +1,11 @@
+
+
+def solveMeFirst(a,b):
+    return a+b
+	# Hint: Type return a+b below
+
+
+num1 = int(input())
+num2 = int(input())
+res = solveMeFirst(num1,num2)
+print(res)
